requirements.txt
Flask==2.0.1
dnspython==2.1.0
idna==3.2
gunicorn==20.1.0